Add Fix code after moving components
