+++ Week-1/src/components/Layout.js
@@ -2,7 +2,7 @@ import React  from 'react'
 import { Route, Switch, NavLink } from 'react-router-dom'
 import { mainPath, aboutPath, productsPath } from '~/src/routes/helpers.js'
 import CartContainer from '~/src/containers/CartContainer.js'
-import Basket from '~/src/ProductCard/Basket.js'
+import Basket from '~/src/components/ProductCard/Basket.js'
 
 // Views
 import routes from '~/src/routes/index.js'
@@ -13,17 +13,17 @@ const routeWithSubroutes = (route, key) =>(
 const Layout = () =>(
   <CartContainer >
     <Basket />
-    <ul>
-      <li>
-        <NavLink to={ mainPath() }>Home</NavLink>
-      </li>
-      <li>
-        <NavLink to={ productsPath() }>Products</NavLink>
-      </li>
-      <li>
-        <NavLink to={ aboutPath() }>About</NavLink>
-      </li>
-    </ul>
+      <ul>
+        <li>
+          <NavLink to={ mainPath() }>Home</NavLink>
+        </li>
+        <li>
+          <NavLink to={ productsPath() }>Products</NavLink>
+        </li>
+        <li>
+          <NavLink to={ aboutPath() }>About</NavLink>
+        </li>
+      </ul>
     <Switch>
       { routes.map((route, key) => routeWithSubroutes(route, key))}
     </Switch>
